Fix circular dependency between aiModelRegistryService and usageTrackingService
- Move IAIModelRegistryService decorator/interface to aiModelRegistryTypes.ts
- Move IUsageTrackingService decorator/interface to usageTrackingTypes.ts
- Replace direct @IUsageTrackingService injection with lazy resolution via @IInstantiationService
- Update 3 test files for new constructor signature

Closes #110
Built by AINative Dev Team
