+++ ainative-studio/src/vs/workbench/contrib/ainative/test/common/authIntegration.test.ts
@@ -205,10 +205,18 @@ suite('Comprehensive Integration Tests - Issue #47 AINative Authentication', () 
 			storageService
 		));
 
+		const mockInstantiationService = {
+			_serviceBrand: undefined,
+			invokeFunction: (fn: any) => fn({ get: () => usageTracking }),
+			createInstance: () => { throw new Error('Not implemented'); },
+			createChild: () => { throw new Error('Not implemented'); },
+			dispose: () => { }
+		};
+
 		modelRegistry = disposables.add(new AIModelRegistryService(
 			authService,
 			storageService,
-			usageTracking
+			mockInstantiationService as any
 		));
 
 		// Update cross-references
